refs #79: moved sidebar to separate class
