+++ CMakeLists.txt
@@ -64,7 +64,7 @@ set(sources
 	${source_dir}/helpers/painter.cpp
 	${source_dir}/helpers/boundingbox.cpp
 	${source_dir}/helpers/positionGenerator.cpp
-)
+	${source_dir}/sidebar.cpp)
 
 set(headers
 	${source_dir}/mosaic.hpp
@@ -80,7 +80,7 @@ set(headers
 	${source_dir}/helpers/painter.hpp
 	${source_dir}/helpers/boundingbox.hpp
 	${source_dir}/helpers/positionGenerator.hpp
-)
+	${source_dir}/sidebar.hpp)
 
 # target and required libs
